chore: 컴포넌트 이름 대문자 변경을 위한 작업
